refactoring - move StateChanges to separate class
