+++ src/main/java/com/wavesplatform/wavesj/transactions/InvokeScriptTransactionStCh.java
@@ -3,11 +3,8 @@ package com.wavesplatform.wavesj.transactions;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.wavesplatform.wavesj.ByteString;
-import com.wavesplatform.wavesj.DataEntry;
 import com.wavesplatform.wavesj.PublicKeyAccount;
-import com.wavesplatform.wavesj.Transfer;
 
-import java.util.Collection;
 import java.util.List;
 
 /**
@@ -37,42 +34,5 @@ public class InvokeScriptTransactionStCh extends InvokeScriptTransaction {
 
     public StateChanges getStateChanges() {
         return stateChanges;
-    }
-
-    public static class StateChanges {
-        private Collection<DataEntry<?>> data;
-        private Collection<OutTransfer> transfers;
-
-        @JsonCreator
-        StateChanges(@JsonProperty("data") Collection<DataEntry<?>> data,
-                     @JsonProperty("transfers") Collection<OutTransfer> transfers) {
-            this.data = data;
-            this.transfers = transfers;
-        }
-
-        public Collection<DataEntry<?>> getData() {
-            return data;
-        }
-
-        public Collection<OutTransfer> getTransfers() {
-            return transfers;
-        }
-    }
-
-    public static class OutTransfer extends Transfer {
-
-        private String assetId;
-
-        @JsonCreator
-        OutTransfer(@JsonProperty("address") String recipient,
-                    @JsonProperty("amount") long amount,
-                    @JsonProperty("asset") String assetId) {
-            super(recipient, amount);
-            this.assetId = assetId;
-        }
-
-        public String getAssetId() {
-            return assetId;
-        }
     }
 }
